requirements.txt
sentence-transformers==2.2.2
langchain
langchain-community
langchain-openai
langchain-experimental
langchain-pinecone
langchain-huggingface
langchain_text_splitters
pinecone[grpc]
flask
pypdf
python-dotenv
sentence-transformers
-e .
